Create BTC network utils
